Move timelimit decorator to its own module
Also remove unused patch_max_test_time utility.
