+++ corehq/util/tests/test_utils.py
@@ -1,33 +1,8 @@
-import re
-import time
-
 from unittest import TestCase
 
-from testil import assert_raises, eq
+from testil import eq
 
-from ..test_utils import disable_quickcache, generate_cases, timelimit
-
-
-def test_timelimit_pass():
-    @timelimit(0.001)
-    def addone(x):
-        return x + 1
-    eq(addone(x=1), 2)
-
-
-def test_timelimit_fail():
-    @timelimit(0.0001)
-    def sleeper():
-        time.sleep(0.001)
-    with assert_raises(AssertionError, msg=re.compile("sleeper took too long")):
-        sleeper()
-
-
-def test_timelimit_default():
-    @timelimit
-    def double(x):
-        return x * 2
-    eq(double(2), 4)
+from ..test_utils import disable_quickcache, generate_cases
 
 
 def test_disable_quickcache():
